Float Speeds and Blue Towers 16/10/2023 8:31PM

+++ enemy.cpp
@@ -26,7 +26,7 @@ enemy::enemy() {
 
 }
 
-enemy::enemy(sf::Texture texture, std::string name, float x, float y, int health, int speed) {
+enemy::enemy(sf::Texture texture, std::string name, float x, float y, int health, float speed) {
 
     // initialise attributes for constructor
     this->name = name;
@@ -45,6 +45,6 @@ enemy::enemy(sf::Texture texture, std::string name, float x, float y, int health
 
 void enemy::take_damage(int damage) {this->health -= damage;}
 
-int enemy::get_speed() {return this->speed;}
+float enemy::get_speed() {return this->speed;}
 
 int enemy::get_health() {return this->health;}
